grpc尝试成功！

+++ model/User.go
@@ -12,9 +12,9 @@ type User struct {
 	FollowerCount   int64  `json:"follower_count,omitempty"`
 	IsFollow        bool   `json:"is_follow,omitempty" gorm:"-"`
 	Password        string `json:"password" gorm:"size:255,not null"`
-	WorkCount       int    `json:"work_count" gorm:"default:0"`
-	TotalFavorited  int    `json:"total_favorited" gorm:"default:0"`
-	FavoriteCount   int    `json:"favorite_count" gorm:"default:0"`
+	WorkCount       int64  `json:"work_count" gorm:"default:0"`
+	TotalFavorited  int64  `json:"total_favorited" gorm:"default:0"`
+	FavoriteCount   int64  `json:"favorite_count" gorm:"default:0"`
 }
 
 type RespUser struct {
@@ -26,7 +26,7 @@ type RespUser struct {
 	FollowCount     int64  `json:"follow_count,omitempty"`
 	FollowerCount   int64  `json:"follower_count,omitempty"`
 	IsFollow        bool   `json:"is_follow,omitempty"`
-	WorkCount       int    `json:"work_count"`
-	TotalFavorited  int    `json:"total_favorited"`
-	FavoriteCount   int    `json:"favorite_count"`
+	WorkCount       int64  `json:"work_count"`
+	TotalFavorited  int64  `json:"total_favorited"`
+	FavoriteCount   int64  `json:"favorite_count"`
 }
